Extract reconciliation template logic into separate class
