+++ lib/reconciliation.rb
@@ -1,3 +1,4 @@
 require_relative './reconciliation/interface'
 require_relative './reconciliation/fuzzer'
 require_relative './reconciliation/legacy_csv'
+require_relative './reconciliation/template'
