+++ templates/reconciliation.html.erb
@@ -34,12 +34,6 @@
 
     <div class="pairings"></div>
 
-    <!--
-    <a href="#" id="generate-csv">Generate CSV</a>
-    <p>Paste this CSV into <code><%= csv_file %></code></p>
-    <textarea id="csv-output" rows="10"></textarea>
-    -->
-
     <script type="text/html" id="pairing">
         <div class="pairing">
             <div class="row">
